+++ src/Renderer/Mesh.h
@@ -22,20 +22,20 @@ class Mesh
 {
 public:
     Mesh();
-    Mesh(Vertex* vertices, uint32_t vertexCount, uint32_t* indices, uint32_t indexCount);
+    Mesh(Vertex* vertices, size_t vertexCount, uint32_t* indices, size_t indexCount);
     ~Mesh();
     void bind() const;
     void unbind() const;
     void draw() const;
 protected:
     Vertex* m_Vertices;
-    uint32_t m_VertexCount;
+    size_t m_VertexCount;
     uint32_t* m_Indices;
-    uint32_t m_TriangleCount;
+    size_t m_TriangleCount;
     uint32_t m_VaoId;
     uint32_t m_IboId;
 
-    void setup(Vertex* vertices, uint32_t vertexCount, uint32_t* indices, uint32_t indexCount);
+    void setup(Vertex* vertices, size_t vertexCount, uint32_t* indices, size_t indexCount);
 };
 
 #endif // MESH_H
